Test Suite for STLCC.Util.Nat, changes in STLCC.Util.*
A couple of utility functions and types added to STLCC.Util.Nat and
STLCC.Util.Vec.

In particular: Utilities for working with existential quantification
over Nats and a NatLTE type + existentials for that.

Also fixes a bug in the arbitrary instance for Fin, and adjusts the
types of a few STLCC.Util.Vec functions.